Release v3.45B1
// Implementation of PID improvements: Derivative low-pass filter, Anti-windup, limit ITerm, reset ITerm
// Implementation of TESTER_IF to report internal time via Serial I/F
// Improved heading in External HDM mode

+++ ActuatorController.h
@@ -54,7 +54,7 @@ protected:
 	int cal_FBK_move(e_dir dir);
 
 	//VIRTUAL ACTUATOR status update
-	float compute_VA (void);
+	int compute_VA (void);
 
 private:
 	e_dir _currentDirection= EXTEND; // EXTEND or RETRACT
